Fix "sbuild --extra-package" usage
This is really some bizarre edge case in APT, but I haven't been able to
reproduce it consistently with a smaller reproducer than fully
re-building several Go packages (golang-github-jinzhu-inflection, then
golang-gopkg-pg.v5 with --extra-package), so I don't feel like we even
have enough information yet to bug the APT maintainers with it.

This simply removes the APT configuration from the "sbuild" variants
which causes the bug to exhibit.