Fix APT comment syntax for older APT versions (instead of just not configuring them)
It turns out that support for `apt.conf` comments of the style `# xxx` was added in 0.7.22 (https://salsa.debian.org/apt-team/apt/commit/81e9789b12374073e848c73c79e235f82c14df44), and that I was incorrect in simply assuming that these configuration options were making this script blow up on really old APT versions!

This adds (and uses) a script that will rewrite `# ...` to `// ...` for APT versions that are old enough to need it.  I've tested it successfully as far back as Debian Potato!