Add workaround for EOL suites with expired GPG keys
For the sake of EOL releases (whose archive keys have often expired), we need a fake `gpgv` substitute that will essentially ignore *just* key expiration.

(So we get *some* signature validation instead of using something like `--allow-unauthenticated` or `--force-yes` which disable security entirely instead.)

This workaround *only* gets applied when we're doing an EOL build, and only when the APT version is new enough to need it (older APT versions didn't verify that).