Fix EOL detection/assumptions via more checking and `distro-info-data`
This fixes `debuerreotype-debian-sources-list` to output into the final file the correct URL based on where it detected the given suite, regardless of the EOL state (instead of always setting the final state to archive.debian.org).

Also, `examples/debian-all.sh` is updated to utilize `distro-info-data` to get more exact EOL dates for each release, thus allowing us to detect that Jessie is EOL even though it's still on the security mirror.